Fix check incoming message length (#1463)
