--- .appveyor.yml
@@ -1,27 +1,0 @@
-os: Visual Studio 2015
-
-environment:
-  # Workaround for the issue https://github.com/briansmith/ring/issues/648
-  CL: "/wd5045"
-  # Stable 64-bit MSVC
-  target: rust-ms
-  # Print full backtrace
-  RUST_BACKTRACE: full
-
-install:
-  - choco install -y %target%
-  - choco install -y protoc
-  - refreshenv
-  - rustc -vV
-  - cargo -vV
-
-build: false
-
-test_script:
-  - cargo test --tests --all
-  # Doc tests are disabled due to appveyor total timeout. 
-  # - cargo test --doc
-
-cache:
-  - '%USERPROFILE%\.cargo -> Cargo.toml'
-  - target -> Cargo.toml
